refactor: create interactionCreate event listener and changed some listeners functions nomenclature
