+++ src/events/utility/customGongo.ts
@@ -1,8 +1,8 @@
 import { Client } from 'discord.js';
-import { playGongo } from '../../listeners/events/customGongo';
+import { onCustomGongo } from '../../listeners/events/customGongo';
 
 export const name = "gongo";
 
 export async function execute(client: Client): Promise<void> {
-  await playGongo(client)
+  await onCustomGongo(client)
 }
