+++ src/listeners/events/interactionCreate/index.ts
@@ -1,0 +1,8 @@
+import { Client, Interaction } from 'discord.js';
+import { handleChatInputCommand } from './handleChatInputCommand';
+
+export async function onInteractionCreate(interaction: Interaction, client: Client): Promise<void> {
+	if (interaction.isChatInputCommand()) {
+		handleChatInputCommand(interaction)
+	}
+}
